orderbook struct and implementation with add order and print

+++ src/orderbook/types.rs
@@ -1,6 +1,6 @@
 use std::time::SystemTime;
 
-pub type Price = f64;
+pub type Price = u64;
 pub type Quantity = u64;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
